Process the missing List of Nested Objects
findSubtypes uses ExtensionLists to determine derived classes.  This is
critcial to any operation that uses anything other than Java classes as
the required object type.  Also some required classes were not
loading/failing right away due to missing Jenkins.  The sort of
convoluted way to fix this was to introduce mocking.  Now, there's a
Jenkins "object" that is barebones enough to handle the necessary
methods to load the missing steps.